gtaxl: Move the proprietary prebuilt libaudioroute.so to system
It seems that the system libaudioroute.so built from sources now takes
precedence over the vendor libaudioroute.so. Since the system
libaudioroute.so doesn't have a audio_route_get_mixer_ctl symbol, the
proprietary audio HAL fails to load:
05-01 09:17:10.002  2610  2610 E vndksupport: Could not load /vendor/lib/hw/audio.primary.universal7870.so from default namespace: dlopen failed: cannot locate symbol "audio_route_get_mixer_ctl" referenced by "/system/vendor/lib/hw/audio.primary.universal7870.so"....
05-01 09:17:10.002  2610  2610 E HAL     : load: module=/vendor/lib/hw/audio.primary.universal7870.so
05-01 09:17:10.002  2610  2610 E HAL     : unknown
05-01 09:17:10.002  2610  2610 E DevicesFactoryHAL: loadAudioInterface couldn't load audio hw module audio.primary (Invalid argument)
With this, the proprietary prebuilt libaudioroute.so in vendor needs to be
moved to system to get it working again.

Change-Id: Ia97c144c4951a066876a456338a46ff1692eede5